Update dependencies and AndroidManifest for USB support
This commit updates the `pubspec.yaml` to include the `usb_serial` package version 0.5.2 and downgrades the `injectable` package to version 2.3.2. Additionally, it modifies the `AndroidManifest.xml` to add USB device configuration and permissions, enhancing the app's capability to interact with USB devices. The macOS configuration files are also updated to include necessary pod support.