Increment version to 1.0.0+12 and enhance DeviceMonitor layout with improved scrolling and data handling
This commit updates the app version to 1.0.0+12 in `pubspec.yaml` and refines the `DeviceMonitor` widget by wrapping its content in a `SingleChildScrollView` for better scrolling behavior. Additionally, a new `ScrollController` is introduced to manage the main scroll view, improving the user experience when interacting with received data and logs. The Android data source is also updated to simplify baud rate configuration for enhanced communication reliability.